mini-03/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.1
sentence-transformers==2.3.1
torch>=2.0.0
numpy
scikit-learn
joblib
wikipedia
